backend_server/requirements.txt
flask
flask-cors
flask-restful
flask-sqlalchemy
flask-pydantic
flask-bcrypt
psycopg2-binary
redis[hiredis]
jwt
python-dotenv
uuid